requirements.txt
streamlit==1.41.1
kiwipiepy==0.17.0
rank-bm25==0.2.2
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
numpy==1.26.4
pandas==2.2.3
markdown==3.7
openpyxl==3.1.5
